SystemComplete

+++ src/main/java/net/brian/islandcore/crop/listener/CropBreakListener.java
@@ -7,9 +7,14 @@ import net.brian.islandcore.common.objects.IslandLogger;
 import net.brian.islandcore.crop.IslandCropService;
 import net.brian.islandcore.crop.objects.ActiveCrop;
 import net.brian.islandcore.crop.objects.IslandCropProfile;
+import org.bukkit.Material;
+import org.bukkit.block.Block;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.event.block.BlockFromToEvent;
+import org.bukkit.event.block.BlockPhysicsEvent;
+import org.bukkit.event.block.BlockPistonExtendEvent;
 
 @Component
 public class CropBreakListener implements Listener {
@@ -23,8 +28,10 @@ public class CropBreakListener implements Listener {
         if(cropService.isCrop(event.getBlock())){
             IslandCropProfile profile = cropService.getProfileFromCrop(event.getBlock());
             ActiveCrop activeCrop = profile.getCrop(event.getBlock());
-            profile.remove(event.getBlock());
-            activeCrop.drop();
+            if(activeCrop != null){
+                profile.remove(event.getBlock());
+                activeCrop.drop();
+            }
         }
     }
 
@@ -33,8 +40,41 @@ public class CropBreakListener implements Listener {
         if(cropService.isCrop(event.getBlock())){
             IslandCropProfile profile = cropService.getProfileFromCrop(event.getBlock());
             ActiveCrop activeCrop = profile.getCrop(event.getBlock());
-            profile.remove(event.getBlock());
-            activeCrop.drop();
+            if(activeCrop != null) {
+                profile.remove(event.getBlock());
+                activeCrop.drop();
+            }
         }
     }
+
+    @EventHandler
+    public void onWater(BlockFromToEvent event){
+        if(cropService.isCrop(event.getToBlock())){
+            Block block = event.getToBlock();
+            IslandCropProfile profile = cropService.getProfileFromCrop(block);
+            ActiveCrop activeCrop = profile.getCrop(block);
+            if(activeCrop != null) {
+                profile.remove(block);
+                activeCrop.drop();
+            }
+        }
+    }
+
+    @EventHandler
+    public void onPiston(BlockPistonExtendEvent event){
+        for(Block block:event.getBlocks()){
+            if(cropService.isCrop(block)){
+                IslandCropProfile profile = cropService.getProfileFromCrop(block);
+                ActiveCrop activeCrop = profile.getCrop(block);
+                if(activeCrop != null) {
+                    profile.remove(block);
+                    activeCrop.drop();
+                    if(activeCrop.getStage() == activeCrop.getCropType().getMaxStage()){
+                        profile.addCount(activeCrop.getCropType().getId());
+                    }
+                }
+            }
+        }
+    }
+
 }
